assistant/pyproject.toml
[project]
name = "assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "openai>=1.98.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.99",
    "pywin32>=311",
    "speechrecognition>=3.14.3",
]


